NEW Apply versioned filters to reading state
This is useful for augmenting methods which return a DataObject
rather than a DataList, with their own internal querying logic.
The most common example for this is SiteTree::get_by_link().

Note that I've split out the argument validation rather than duplicating it,
and throw exceptions for the bits which aren't supported in "reading mode"
(namely "state" and "latest"versions")